requirements.txt
aiohttp==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
discord.py==1.5.1
ffmpeg==1.4
fuzzywuzzy==0.18.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
multidict==4.7.6
opuslib==3.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
python-Levenshtein==0.12.0
pytz==2020.4
requests==2.25.0
rich==9.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tqdm==4.54.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
yarl==1.5.1
youtube-dl==2020.12.2
